#23 Test cases fix for workflows
